\name{dma-package}
\alias{dma-package}
\docType{package}
\title{
Dynamic model averaging
}
\description{
This package implements dynamic Bayesian model averaging as described for continuous outcomes in Raftery et al. (2010, Technometrics) and for binary outcomes in McCormick et al. (2011, Biometrics).
}
\details{
\tabular{ll}{
Package: \tab dma\cr
Type: \tab Package\cr
Version: \tab 1.2-0\cr
Date: \tab 2013-05-2\cr
License: \tab GPL2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Tyler H. McCormick, Adrian Raftery, David Madigan

Maintainer: Hana Sevcikova <hanas@uw.edu>
}
\references{
McCormick, T.M., Raftery, A.E., Madigan, D. and Burd, R.S. (2011) "Dynamic Logistic Regression and 
Dynamic Model Averaging for Binary Classification." Biometrics, 66:1162-1173.

Raftery, A.E., Karny, M., and Ettler, P. (2010). Online Prediction Under Model 
Uncertainty Via Dynamic Model Averaging: Application to a Cold Rolling Mill. Technometrics 52:52-66.
}
\keyword{dma}


