% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dma_control.R
\name{dma_control}
\alias{dma_control}
\title{Control parameters for dma}
\usage{
dma_control(
  crossfit_folds = 10L,
  zprime_folds = 1L,
  noise_dim = 5L,
  hidden_dim = 100L,
  num_layer = 3L,
  num_epochs = 500L,
  lr = 0.001,
  riesz_epochs = 100L,
  riesz_lr = 0.01,
  batch_size = 64L,
  riesz_batch_size = 64L,
  device = "cpu"
)
}
\arguments{
\item{crossfit_folds}{[\code{numeric(1)}]\cr Number of crossfit folds.}

\item{zprime_folds}{[\code{numeric(1)}]\cr Number of folds for calculating Z'.}

\item{noise_dim}{[\code{numeric(1)}]\cr Engression noise dimension.}

\item{hidden_dim}{[\code{numeric(1)}]\cr NN hidden layer dimension.}

\item{num_layer}{[\code{numeric(1)}]\cr Number of NN layers for engression.}

\item{num_epochs}{[\code{numeric(1)}]\cr Training epochs for outcome regressions.}

\item{lr}{[\code{numeric(1)}]\cr Learning rate for outcome regressions.}

\item{riesz_epochs}{[\code{numeric(1)}]\cr Training epochs for Riesz representer.}

\item{riesz_lr}{[\code{numeric(1)}]\cr Learning rate for Riesz representer.}

\item{batch_size}{[\code{numeric(1)}]\cr Batch size for outcome regressions.}

\item{riesz_batch_size}{[\code{numeric(1)}]\cr Batch size for Riesz representer.}

\item{device}{[\code{character(1)}]\cr Torch device: "cpu", "cuda", or "mps".}
}
\value{
A list of control parameters.
}
\description{
Control parameters for dma
}
