% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{new_dma_result}
\alias{new_dma_result}
\title{dma result object}
\usage{
new_dma_result(
  estimates,
  outcome_reg,
  alpha_n,
  alpha_r,
  models_y,
  folds,
  vars,
  data,
  call,
  effect,
  d0,
  d1
)
}
\arguments{
\item{estimates}{A list of effect estimates.}

\item{outcome_reg}{Outcome regression predictions.}

\item{alpha_n}{Natural density ratio estimates.}

\item{alpha_r}{Randomized density ratio estimates.}

\item{models_y}{List of trained engression models.}

\item{folds}{List of cross-fit folds.}

\item{vars}{Object of class dma_vars.}

\item{data}{Data frame used for estimation.}

\item{call}{The matched call.}

\item{effect}{The estimated effect type.}
}
\value{
An object of class \code{dma_result}.
}
\description{
dma result object
}
