% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential_module.R
\name{sequential_module}
\alias{sequential_module}
\title{Sequential neural network module function factory}
\usage{
sequential_module(layers = 1, hidden = 20, dropout = 0.1)
}
\arguments{
\item{layers}{[numeric(1)]\cr Number of hidden layers.}

\item{hidden}{[numeric(1)]\cr Number of hidden units.}

\item{dropout}{[numeric(1)]\cr Dropout rate.}
}
\value{
A function that takes input dimension and returns a torch nn_module.
}
\description{
Creates a factory function that returns a sequential neural network module
with the specified architecture. Used for Riesz representer estimation.
}
